organize project, add extras button
