+++ Source/EzOutfit/EzOutfit.csproj
@@ -66,9 +66,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Entry.cs" />
+    <Compile Include="UI\CreateFromPawnButton.cs" />
+    <Compile Include="UI\ExtrasButton.cs" />
     <Compile Include="OutfitDialoguePatch.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="UI\ExtrasMenu.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
