+++ Source/EzOutfit/UI/ExtrasMenu.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+public static class ExtrasMenu
+{
+  public static void DoMenu()
+  {
+    
+  }
+}
